Add files via upload
These are the files in which I've explored the data, got an understanding of various features, and on the basis of that created 3 MySQL tables one storing the unique country names and their ISO codes (abbreviations), 2nd table storing various electrical energy data, and 3rd table storing the same with respect to continents. In both the tables, there are two primary keys. The ISO code and the year for which the data was recorded (2000-2023).